Rename Screenshot (13).png to Detect vehicle and check whether it is positioned in the center.png
Vehicle Detection with YOLO (You Only Look Once) and OpenCV: This project utilizes the YOLO object detection algorithm,
implemented using OpenCV, to detect and highlight vehicles in videos, images, and even URLs. The system identifies vehicles, checks if they are in the center of the frame, and displays the results.