Add files via upload
🚀 Excited to share my latest project on Emotion Detection using AI! 🎯

Using hashtag#TensorFlow, hashtag#OpenCV, and a custom-trained model, I built a system that detects human emotions from images with incredible accuracy! 😊😡😱 Whether it’s happiness, anger, surprise, or sadness, this model can read it all in real time and display it with a personalized green frame! 🖼️
🔍 Real-time emotion insights can have amazing applications in fields like mental health, customer feedback analysis, and human-computer interaction. 🧠💬

Check it out and feel free to reach out for collaboration or feedback! 🙌